+++ prisma/postgresql-migrations/20241011085129_create_pusher_table/migration.sql
@@ -1,0 +1,22 @@
+-- CreateTable
+CREATE TABLE "Pusher" (
+    "id" TEXT NOT NULL,
+    "enabled" BOOLEAN NOT NULL DEFAULT false,
+    "appId" VARCHAR(100) NOT NULL,
+    "key" VARCHAR(100) NOT NULL,
+    "secret" VARCHAR(100) NOT NULL,
+    "cluster" VARCHAR(100) NOT NULL,
+    "useTLS" BOOLEAN NOT NULL DEFAULT false,
+    "events" JSONB NOT NULL,
+    "createdAt" TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" TIMESTAMP NOT NULL,
+    "instanceId" TEXT NOT NULL,
+
+    CONSTRAINT "Pusher_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Pusher_instanceId_key" ON "Pusher"("instanceId");
+
+-- AddForeignKey
+ALTER TABLE "Pusher" ADD CONSTRAINT "Pusher_instanceId_fkey" FOREIGN KEY ("instanceId") REFERENCES "Instance"("id") ON DELETE CASCADE ON UPDATE CASCADE;