+++ src/api/integrations/event/event.router.ts
@@ -1,3 +1,4 @@
+import { PusherRouter } from '@api/integrations/event/pusher/pusher.router';
 import { RabbitmqRouter } from '@api/integrations/event/rabbitmq/rabbitmq.router';
 import { SqsRouter } from '@api/integrations/event/sqs/sqs.router';
 import { WebhookRouter } from '@api/integrations/event/webhook/webhook.router';
@@ -13,6 +14,7 @@ export class EventRouter {
     this.router.use('/webhook', new WebhookRouter(configService, ...guards).router);
     this.router.use('/websocket', new WebsocketRouter(...guards).router);
     this.router.use('/rabbitmq', new RabbitmqRouter(...guards).router);
+    this.router.use('/pusher', new PusherRouter(...guards).router);
     this.router.use('/sqs', new SqsRouter(...guards).router);
   }
 }