+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fake Call function
 * Added unreadMessages to chats
+* Pusher event integration
 
 ### Fixed